backend/dynamic_nft_avatars/pyproject.toml
[tool.poetry]
name = "donft"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
aiographql-client = "^1.0.2"
numpy = "^1.21.4"
python-multipart = "^0.0.5"
environs = "^9.3.5"
imageio = "2.3.0"
pandas = "0.25.3"
torch = "1.8.0"
torchvision = "0.2.1"
scipy = "^1.7.3"
scikit-image = "^0.18.3"
scikit-learn = "^1.0.1"
matplotlib = "^3.5.0"
tqdm = "^4.62.3"
face-alignment = "^1.3.5"
web3 = "^5.28.0"
tensorflow = "2.7.0"
tensorflow-hub = "0.12.0"
Pillow = "7.1.2"
facenet-pytorch = "^2.5.2"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"
flake8 = "^4.0.1"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
